File separation: inference.py and data_engineering.py removed from notebooks.ipynb in notebooks folder to src/pipeline folder. Model:  model folder deleted and trainded model added in models/trained folder
