--- model/history/history-v1.json
@@ -1,124 +1,0 @@
-{
-  "train_loss": [
-    0.019119790415796967,
-    0.01190336500811908,
-    0.011991744107670255,
-    0.011464977636933327,
-    0.011419750649688972,
-    0.011406602923654847,
-    0.011398963619851404,
-    0.011390233640041617,
-    0.011388840698119667,
-    0.011384611787895361
-  ],
-  "val_loss": [
-    0.010783518515527248,
-    0.010388065092265607,
-    0.010311294347047806,
-    0.010301003605127335,
-    0.010292632691562175,
-    0.010293550863862038,
-    0.010265091806650162,
-    0.010272633470594882,
-    0.010263303965330125,
-    0.010251813717186451
-  ],
-  "metrics": {
-    "mse": [
-      0.01078351825417485,
-      0.010388065063598333,
-      0.010311294421808271,
-      0.01030100385632977,
-      0.010292632799610147,
-      0.010293550621136092,
-      0.010265091782675881,
-      0.010272633594031504,
-      0.010263303913234267,
-      0.010251813766699343
-    ],
-    "mae": [
-      0.06405167507007718,
-      0.060889678290113804,
-      0.060047472898149865,
-      0.059919006936252114,
-      0.05977057251380757,
-      0.059765334874391556,
-      0.05971749827964232,
-      0.05976996842306107,
-      0.059700015929993244,
-      0.05965995396254584
-    ],
-    "snr_db": [
-      -0.9646210084963365,
-      -0.22780197209740613,
-      -0.056077640839670616,
-      -0.049554264826001324,
-      -0.03423514886568013,
-      -0.03040719579016909,
-      -0.02378969300714906,
-      -0.03501319831460255,
-      -0.022991960209874505,
-      -0.02094439930258286
-    ],
-    "si_sdr": [
-      -45.048100721915226,
-      -45.43236315489458,
-      -44.74880259778335,
-      -45.50627009772435,
-      -45.2170132993283,
-      -43.143760579392,
-      -41.9077460177083,
-      -44.24197024050293,
-      -42.102599204152256,
-      -41.92852413067469
-    ],
-    "sdr": [
-      -45.048100721915226,
-      -45.43236315489458,
-      -44.74880259778335,
-      -45.50627009772435,
-      -45.2170132993283,
-      -43.143760579392,
-      -41.9077460177083,
-      -44.24197024050293,
-      -42.102599204152256,
-      -41.92852413067469
-    ],
-    "sir": [
-      -45.048100721915226,
-      -45.43236315489458,
-      -44.74880259778335,
-      -45.50627009772435,
-      -45.2170132993283,
-      -43.143760579392,
-      -41.9077460177083,
-      -44.24197024050293,
-      -42.102599204152256,
-      -41.92852413067469
-    ],
-    "sar": [
-      91.10468206186997,
-      85.3323289118004,
-      82.28384823801065,
-      80.23681157651995,
-      78.8392016357401,
-      78.63855580019313,
-      77.44193103013325,
-      77.48308475151819,
-      76.62200326894329,
-      75.12525520132247
-    ],
-    "isr": [
-      -45.048100721915226,
-      -45.43236315489458,
-      -44.74880259778335,
-      -45.50627009772435,
-      -45.2170132993283,
-      -43.143760579392,
-      -41.9077460177083,
-      -44.24197024050293,
-      -42.102599204152256,
-      -41.92852413067469
-    ]
-  }
-}
