--- model/history/history-v2.json
@@ -1,76 +1,0 @@
-{
-  "train_loss": [
-    0.08015665679445697,
-    0.001079055974777374,
-    0.0005094416559182314,
-    0.0003422988429924266,
-    0.0002558748313782012,
-    0.00020793967058933856,
-    0.00017025926596640298,
-    0.0001494870370758387,
-    0.0001362266765985017,
-    0.00012496448164003797
-  ],
-  "val_loss": [
-    0.002062115631997585,
-    0.0011010600486770273,
-    0.0008176200906746089,
-    0.0006084945693146437,
-    0.0005237651173956692,
-    0.00040873581892810764,
-    0.00036564808688126503,
-    0.0003296862647403032,
-    0.000310732057550922,
-    0.00029731600952800365
-  ],
-  "metrics": {
-    "mse": [
-      0.00780771931167692,
-      0.009318339102901519,
-      0.009796246467158199,
-      0.00940596596337855,
-      0.009499930660240352,
-      0.009310353035107254,
-      0.009319571149535476,
-      0.009331781882792711,
-      0.009335960960015655,
-      0.009351393883116543
-    ],
-    "mae": [
-      0.0666709529235959,
-      0.07712731659412383,
-      0.07916171066462993,
-      0.07734616287052631,
-      0.07730746380984783,
-      0.0768266711384058,
-      0.07699736319482327,
-      0.07714912556111812,
-      0.07717428021132947,
-      0.07723480649292469
-    ],
-    "snr_db": [
-      1.7844986663677989,
-      0.7483060713206292,
-      0.5407234038804357,
-      0.7488301593707264,
-      0.7339243054763127,
-      0.8179443275196124,
-      0.8141027310129492,
-      0.807566345343643,
-      0.8076110593245582,
-      0.8015476273357546
-    ],
-    "si_sdr": [
-      10.788614429431686,
-      11.133012477296319,
-      10.50453961018179,
-      10.620995846665178,
-      10.32297670319676,
-      10.449232560482269,
-      10.506047861318816,
-      10.418680377910784,
-      10.406160710360775,
-      10.37270390598621
-    ]
-  }
-}
