+++ requirements.txt
@@ -1,0 +1,21 @@
+torch>=1.9.0
+torchaudio>=0.9.0
+numpy>=1.21.0
+scipy>=1.7.0
+
+librosa>=0.9.0
+soundfile>=0.10.0
+
+pandas>=1.3.0
+tqdm>=4.62.0
+
+PyYAML>=6.0
+
+Flask>=2.0.0
+Werkzeug>=2.0.0
+
+mlflow>=1.20.0
+
+pytest>=6.0.0
+jupyter>=1.0.0
+matplotlib>=3.4.0
