+++ src/pipeline/model/crop.py
@@ -1,0 +1,1 @@
+# crop.py
