+++ src/pipeline/model/utils.py
@@ -1,0 +1,2 @@
+# utils.py
+# Utility functions for Wave-U-Net model
