+++ src/pipeline/model_pipeline.py
@@ -1,0 +1,1 @@
+# model_pipeline.py
